+++ src/main/projectData.ts
@@ -1,19 +1,20 @@
 import { app } from "electron";
-import { ListEntry } from "../../core/pub-sort";
+import { IssueData, IssueDate } from "../../core/issue_data";
 import { APPID } from "./utils";
 import { DEFAULT_FILTER_OPTIONS, DisplayOptions } from "./displayOptions";
 import { loadList, Path } from "../../core/load";
+import { generateCsv, mkConfig, asString } from "export-to-csv";
 import fs from "fs";
 
 export class ProjectData {
   header: { appID: typeof APPID; version: string };
   meta: { characterName: string; title: string; opts: DisplayOptions };
-  list: ListEntry[];
+  list: IssueData[];
 
   private constructor(
     header: { appID: typeof APPID; version: string },
     meta: { characterName: string; title: string; opts: DisplayOptions },
-    list: ListEntry[],
+    list: IssueData[],
   ) {
     this.header = header;
     this.meta = meta;
@@ -41,14 +42,12 @@ export class ProjectData {
   saveAsJSON(path: Path) {
     this.meta.title = path.fileName;
     console.log(this.meta.title);
-    fs.writeFileSync(path.fullPath, JSON.stringify(this), { encoding: "utf8" });
+    fs.writeFileSync(path.fullPath, JSON.stringify(this, null, 2), { encoding: "utf8" });
   }
 
   saveAsMDList(path: Path) {
     let doc = "";
     this.list.map((entry) => {
-      // TODO:need to create the link
-      // TODO: Take the url thing from the renderer and make a URL field on the listentry
       if (this.meta.opts.showDates) {
         doc += `* [${entry.title}](${entry.URL})      ${entry.date.toString()}\n`;
       } else {
@@ -60,22 +59,30 @@ export class ProjectData {
     fs.writeFileSync(path.fullPath, doc, { encoding: "utf8" });
   }
 
+  /**Save the project's list as a CSV */
+  saveAsCSV(path: Path) {
+    // Have to flatten the data because it can't handle nested objects
+    const raw = this.list.map((issue) => {
+      let date: string;
+      if (issue.dateInferred) {
+        date = "MISSING";
+      } else {
+        date = IssueDate.toYYYYMMDD(issue.date);
+      }
+      return { title: issue.title, synopsis: issue.synopsis, date, link: issue.link, URL: issue.URL };
+    });
+    const csvConfig = mkConfig({ useKeysAsHeaders: true, quoteStrings: true });
+    const csv = generateCsv(csvConfig)(raw);
+
+    fs.writeFileSync(path.fullPath, asString(csv), { encoding: "utf8" });
+  }
+
   /**Loads `ProjectData` from a JSON file. Error if the file is not a JSON. */
   static fromJSON(path: Path): ProjectData {
     // Return early with an error if the wrong filetype is passed
     if (!/.jsonc|.json/.test(path.ext)) throw new Error("Incorrect file type.");
     const file = fs.readFileSync(path.fullPath, { encoding: "utf8" });
     const data = JSON.parse(file) as ProjectData;
-
-    // Convert the entries to list entries because right now they don't actually have the class' metadata and functions
-    let entries = [];
-    for (const entry of data.list) {
-      entries.push(
-        new ListEntry(entry.title, entry.synopsis, entry.date.year.toString(), entry.date.month.toString(), entry.date.day.toString()),
-      );
-    }
-
-    data.list = entries;
 
     // TODO: I would prefer not to hard code this but it's not a huge deal
     // Check this is the right app or error
@@ -97,5 +104,5 @@ export class ProjectData {
 export interface SerializedProjectData {
   header: { appID: typeof APPID; version: string };
   meta: { characterName: string; title: string; opts: DisplayOptions };
-  list: ListEntry[];
+  list: IssueData[];
 }